+++ packages/vscode-extension/src/commands/convert.ts
@@ -4,13 +4,18 @@ import { markdownToPretext } from "md2ptx";
 import { pretextOutputChannel } from "../ui";
 import { convertToPretext } from "../importFiles";
 import { processLatexViaUnified } from "@unified-latex/unified-latex";
+import { htmlLike } from "@unified-latex/unified-latex-util-html-like";
+import { getArgsContent } from "@unified-latex/unified-latex-util-arguments";
 import {
   unifiedLatexToPretext,
+  unifiedLatexWrapPars,
   xmlCompilePlugin,
 } from "@unified-latex/unified-latex-to-pretext";
 import { formatPTX } from "../formatter";
 // @ts-ignore
 import { FlexTeXtConvert } from "frankenmarkup";
+import { get } from "http";
+import { Environment, Macro } from "@unified-latex/unified-latex-types";
 
 export function cmdConvertToPretext() {
   console.log("Converting to PreTeXt");
@@ -66,10 +71,46 @@ export function cmdLatexToPretext() {
   }
 }
 
+//const myMacroReplacements = {
+//  myfoo: (node: Macro) => {
+//    console.log("myfoo node is", node.args);
+//    const args = getArgsContent(node);
+//    console.log("args are", args);
+//    return htmlLike({
+//      tag: "myptxfoo",
+//    });
+//  }
+//}
+
+const ptxExtraEnvironmentReplacements = {
+  solution: (node: Environment) => {
+    return htmlLike({
+      tag: "solution",
+      content: node.content,
+    });
+  },
+  answer: (node: Environment) => {
+    return htmlLike({
+      tag: "answer",
+      content: node.content,
+    });
+  },
+  hint: (node: Environment) => {
+    return htmlLike({
+      tag: "hint",
+      content: node.content,
+    });
+  },
+};
+
 function convertWithUnified(text: string) {
   const convert = (value: string) =>
     processLatexViaUnified()
-      .use(unifiedLatexToPretext, { producePretextFragment: true })
+      .use(unifiedLatexToPretext, {
+        producePretextFragment: true,
+        //macroReplacements: myMacroReplacements,
+        environmentReplacements: ptxExtraEnvironmentReplacements,
+      })
       .use(xmlCompilePlugin)
       .processSync({ value });
 
@@ -80,7 +121,6 @@ function convertWithUnified(text: string) {
       console.log(message);
     }
   }
-  console.log("HERE!");
   return convert(text).value as string;
 }
 
@@ -117,9 +157,9 @@ export async function cmdConvertFlextextToPretext() {
     const selection = editor.selection;
     const selectionRange = selection.isEmpty
       ? new Range(
-        editor.document.positionAt(0),
-        editor.document.positionAt(editor.document.getText().length)
-      )
+          editor.document.positionAt(0),
+          editor.document.positionAt(editor.document.getText().length)
+        )
       : new Range(selection.start, selection.end);
     console.log("selectionRange is", selectionRange);
     const initialText = editor.document.getText(selectionRange);
@@ -129,8 +169,8 @@ export async function cmdConvertFlextextToPretext() {
     editor.edit((editbuilder) => {
       editbuilder.replace(selectionRange, newText);
     });
-    // Call default formatter to format the text.
-    commands.executeCommand("editor.action.formatDocument");
+    // Call default formatter to format just the replaced selection.
+    commands.executeCommand("editor.action.formatSelection");
 
     console.log("Formatted text");
   }